backend/requirements.txt
# Backend dependencies for Sign Language Recognition Tool

# Web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Audio processing
speech_recognition==3.10.0
pydub==0.25.1
pyaudio==0.2.11

# Text-to-Speech
gTTS==2.3.2
pyttsx3==2.90

# Google Cloud APIs
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.21.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1

# Machine Learning (for future implementation)
# torch==2.0.1
# torchvision==0.15.2
# opencv-python==4.8.0.76
# numpy==1.24.3
# pandas==2.0.3

# Utilities
requests==2.31.0
python-dotenv==1.0.0